+++ crates/gmsol/src/chainlink/pull_oracle/access_controller_address.rs
@@ -1,5 +1,0 @@
-#[cfg(feature = "devnet")]
-solana_sdk::declare_id!("2k3DsgwBoqrnvXKVvd7jX7aptNxdcRBdcd5HkYsGgbrb");
-
-#[cfg(not(feature = "devnet"))]
-solana_sdk::declare_id!("7mSn5MoBjyRLKoJShgkep8J17ueGG8rYioVAiSg5YWMF");