+++ crates/gmsol/src/discover/mod.rs
@@ -1,5 +1,0 @@
-/// Market discovery.
-pub mod market;
-
-/// Token discovery.
-pub mod token;