+++ crates/gmsol/src/program_ids.rs
@@ -1,1 +1,0 @@
-pub use gmsol_store::ID_CONST as DEFAULT_GMSOL_STORE_ID;