+++ crates/gmsol/src/store/events.rs
@@ -1,1 +1,0 @@
-pub use crate::decode::gmsol::store::GMSOLCPIEvent as StoreCPIEvent;