+++ crates/gmsol/src/switchboard/mod.rs
@@ -1,3 +1,0 @@
-/// Switchboard Pull Oracle.
-#[cfg(feature = "switchboard-pull-oracle")]
-pub mod pull_oracle;