+++ crates/gmsol/src/switchboard/pull_oracle/mod.rs
@@ -1,3 +1,0 @@
-mod pull_oracle_impl;
-
-pub use pull_oracle_impl::{SbPriceUpdates, SwitchboardPullOracle, SwitchcboardPullOracleFactory};