+++ crates/gmsol/src/utils/workarounds/mod.rs
@@ -1,5 +1,0 @@
-/// Workaround for optional accounts.
-pub mod optional;
-
-/// Zero-copy account workaround.
-pub mod zero_copy;