+++ crates/gmsol/tests/integration.rs
@@ -1,2 +1,0 @@
-#[cfg(feature = "integration-test")]
-mod integration_test;