+++ crates/gmsol/tests/integration_test/mod.rs
@@ -1,7 +1,0 @@
-/// Setup.
-pub mod setup;
-
-/// Test Client.
-pub mod test_client;
-
-mod oracle;